Keep Koin and Compose plugins out of MinGW, drop the last Windows-only scaffolding
- Koin.kt moves to nonWindowsMain unchanged from upstream apart from the
  two pieces the Windows client also needs (PeopleInSpaceDatabaseWrapper,
  createHttpClient), which go to a small commonMain file. That removes the
  MinGW NativeModule actual, KoinViewModels.kt, and the Swift call-site
  edits, since the ViewModel factories are back on KoinKt.
- The Compose compiler plugin joins Koin's on the MinGW exclusion list:
  with Compose back inside common it was being applied to
  compileKotlinMingwX64, which has no Compose runtime (Windows CI on
  9f0c21d).
- LocalAppDataStore only ever produced a directory path; App.xaml.cs does
  that inline.
- libs.versions.toml formatting restored to upstream.

Verified on Windows: compileKotlinMingwX64 --rerun-tasks, :common:jvmTest,
:app:compileDebugKotlin, :compose-desktop:compileKotlin, packNuget, clean
restore and 0-warning Release build.
