+++ common/src/commonMain/kotlin/dev/johnoreilly/common/di/Dependencies.kt
@@ -1,0 +1,26 @@
+package dev.johnoreilly.common.di
+
+import app.cash.sqldelight.db.SqlDriver
+import dev.johnoreilly.peopleinspace.db.PeopleInSpaceDatabase
+import io.ktor.client.*
+import io.ktor.client.engine.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.client.plugins.logging.*
+import io.ktor.serialization.kotlinx.json.*
+import kotlinx.serialization.json.Json
+
+// Shared by the Koin graph (nonWindows targets) and the Windows client, which builds its own graph.
+
+class PeopleInSpaceDatabaseWrapper(val driver: SqlDriver, val instance: PeopleInSpaceDatabase)
+
+fun createHttpClient(httpClientEngine: HttpClientEngine, json: Json, enableNetworkLogs: Boolean) = HttpClient(httpClientEngine) {
+    install(ContentNegotiation) {
+        json(json)
+    }
+    if (enableNetworkLogs) {
+        install(Logging) {
+            logger = Logger.DEFAULT
+            level = LogLevel.INFO
+        }
+    }
+}
